Refactor code quality
Changed wildcard imports to individual packages and cleaned up redundant code lines.
