--- src/main/java/Commands/CommandResult.java
@@ -1,9 +1,0 @@
-package Commands;
-
-public class CommandResult {
-    public String commandOutput;
-
-    public CommandResult(String commandOutput) {
-        this.commandOutput = commandOutput;
-    }
-}
