--- src/main/java/Commands/ExitCommand.java
@@ -1,10 +1,0 @@
-package Commands;
-
-public class ExitCommand extends Command{
-    @Override
-    public CommandResult execute() {
-        System.out.println("Exiting application...\n");
-        return new ExitCommandResult("Bye-bye!");
-    }
-
-}
