--- src/main/java/Commands/FindCommand.java
@@ -1,14 +1,0 @@
-package Commands;
-
-public class FindCommand extends Command {
-    String finder;
-
-    public FindCommand(String finder) {
-        this.finder = finder;
-    }
-
-    @Override
-    public CommandResult execute() {
-        return new CommandResult(this.tasks.findTasks(finder));
-    }
-}
