--- src/main/java/Commands/ListCommand.java
@@ -1,9 +1,0 @@
-package Commands;
-
-public class ListCommand extends Command{
-    @Override
-    public CommandResult execute() {
-        String res = tasks.listTasks();
-        return new CommandResult(res);
-    }
-}
