--- src/main/java/Commons/Utils.java
@@ -1,9 +1,0 @@
-package Commons;
-
-public class Utils {
-    public static final String DIVIDER = "____________________________________________________________\n";
-
-    public static void printLine() {
-        System.out.println(DIVIDER);
-    }
-}
