+++ src/main/java/commands/CommandResult.java
@@ -1,0 +1,9 @@
+package commands;
+
+public class CommandResult {
+    public String commandOutput;
+
+    public CommandResult(String commandOutput) {
+        this.commandOutput = commandOutput;
+    }
+}
