+++ src/main/java/commands/ExitCommand.java
@@ -1,0 +1,10 @@
+package commands;
+
+public class ExitCommand extends Command{
+    @Override
+    public CommandResult execute() {
+        System.out.println("Exiting application...\n");
+        return new ExitCommandResult("Bye-bye!");
+    }
+
+}
