+++ src/main/java/commands/ExitCommandResult.java
@@ -1,0 +1,11 @@
+package commands;
+
+public class ExitCommandResult extends CommandResult{
+    public ExitCommandResult(String s) {
+        super(s);
+    }
+
+    public void quit() {
+        System.exit(0);
+    }
+}
