+++ src/main/java/commands/FindCommand.java
@@ -1,0 +1,14 @@
+package commands;
+
+public class FindCommand extends Command {
+    String finder;
+
+    public FindCommand(String finder) {
+        this.finder = finder;
+    }
+
+    @Override
+    public CommandResult execute() {
+        return new CommandResult(this.tasks.findTasks(finder));
+    }
+}
