+++ src/main/java/commands/ListCommand.java
@@ -1,0 +1,9 @@
+package commands;
+
+public class ListCommand extends Command{
+    @Override
+    public CommandResult execute() {
+        String res = tasks.listTasks();
+        return new CommandResult(res);
+    }
+}
