+++ src/main/java/commons/Utils.java
@@ -1,0 +1,9 @@
+package commons;
+
+public class Utils {
+    public static final String DIVIDER = "____________________________________________________________\n";
+
+    public static void printLine() {
+        System.out.println(DIVIDER);
+    }
+}
