+++ src/main/java/exceptions/AikhsuException.java
@@ -1,0 +1,11 @@
+package exceptions;
+
+public class AikhsuException extends Exception{
+    public AikhsuException(String message) {
+        super(message);
+    }
+
+    public AikhsuException(String message, Throwable description) {
+        super(message, description);
+    }
+}
