--- src/test/java/Tasks/TaskListTest.java
@@ -1,17 +1,0 @@
-package Tasks;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class TaskListTest {
-    @Test
-    public void dummyTest(){
-        assertEquals(2, 2);
-    }
-
-    @Test
-    public void anotherDummyTest(){
-        assertEquals(4, 4);
-    }
-}
