+++ src/test/java/tasks/TaskListTest.java
@@ -1,0 +1,17 @@
+package tasks;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class TaskListTest {
+    @Test
+    public void dummyTest(){
+        assertEquals(2, 2);
+    }
+
+    @Test
+    public void anotherDummyTest(){
+        assertEquals(4, 4);
+    }
+}
